requirements.txt
requests==2.23.0
bs4==0.0.1
python-telegram-bot==12.8